New round of cleanup
