+++ app/src/main/java/com/zeapo/pwdstore/di/modules/FragmentModule.java
@@ -1,6 +1,6 @@
 package com.zeapo.pwdstore.di.modules;
 
-import com.zeapo.pwdstore.crypto.fragments.PwgenDialogFragment;
+import com.zeapo.pwdstore.crypto.password.fragments.PwgenDialogFragment;
 
 import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
